requirements.txt
pillow 
screeninfo 
opencv-python 
pywin32 
